Refactor useProfileForm hook, update StepTwo, fix test issues
